Break out raw subscription test to own file and remove panics.
